requirements.txt
pillow>=3.0.0
requests>=2.15.0